// apps/twitter/src/modules/twitter-record/infrastructure/repositories/twitter-record.repository-type.ts
export interface GetFlatChildTreeOfCommentsByRecordIdSqlQueryResult {
  commentedRecordAuthorId: string;
  commentedRecordId: string;
  createdAt: Date;
  id: string;
  text: string;
  twitterRecordImageId: string;
  twitterRecordImageKey: string;
  twitterRecordImageUrl: string;
  updatedAt: Date;
  userId: string;
}